refactor: [MON-279] Wrapper컴포넌트 리팩토링 및 적용
n
