+++ src/pages/general/NotFoundPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Wrapper } from '@components';
 
 const NotFoundPage = () => (
-  <Wrapper whole>
+  <Wrapper>
     <h1>404 Not Found</h1>
   </Wrapper>
 );
