+++ src/styles/index.jsx
@@ -1,0 +1,2 @@
+export { default as mixin } from './mixin';
+export { default as theme } from './theme';
